refactor(coral): Rename useFiltersValues -> useFiltersContext
