+++ coral/src/app/features/approvals/schemas/SchemaApprovals.tsx
@@ -14,7 +14,7 @@ import { SearchTopicFilter } from "src/app/features/components/filters/SearchTop
 import {
   useFiltersContext,
   withFiltersContext,
-} from "src/app/features/components/filters/useFiltersValues";
+} from "src/app/features/components/filters/useFiltersContext";
 import {
   approveSchemaRequest,
   declineSchemaRequest,
