+++ coral/src/app/features/components/filters/EnvironmentFilter.tsx
@@ -1,6 +1,6 @@
 import { NativeSelect, Option } from "@aivenio/aquarium";
 import { useQuery } from "@tanstack/react-query";
-import { useFiltersContext } from "src/app/features/components/filters/useFiltersValues";
+import { useFiltersContext } from "src/app/features/components/filters/useFiltersContext";
 import {
   Environment,
   getAllEnvironmentsForTopicAndAcl,
