+++ coral/src/app/features/topics/browse/BrowseTopics.tsx
@@ -7,7 +7,7 @@ import { SearchTopicFilter } from "src/app/features/components/filters/SearchTop
 import {
   useFiltersContext,
   withFiltersContext,
-} from "src/app/features/components/filters/useFiltersValues";
+} from "src/app/features/components/filters/useFiltersContext";
 import { TableLayout } from "src/app/features/components/layouts/TableLayout";
 import TopicTable from "src/app/features/topics/browse/components/TopicTable";
 import { getTopics } from "src/domain/topic";
